DeepL 6.1.2.6  - fix backup path issue  - fix glossary initialization error message
