+++ DeepLMTProvider/Sdl.Community.DeelLMTProvider/DeepLTellMe/DeepLSettingsAction.cs
@@ -1,16 +1,12 @@
 ﻿using Newtonsoft.Json;
 using Sdl.Community.DeepLMTProvider.Client;
-using Sdl.Community.DeepLMTProvider.Helpers;
-using Sdl.Community.DeepLMTProvider.Helpers.GlossaryReadersWriters;
 using Sdl.Community.DeepLMTProvider.Model;
 using Sdl.Community.DeepLMTProvider.Service;
 using Sdl.Community.DeepLMTProvider.UI;
 using Sdl.Community.DeepLMTProvider.ViewModel;
 using Sdl.TellMe.ProviderApi;
 using Sdl.TranslationStudioAutomation.IntegrationApi;
-using System;
 using System.Drawing;
-using System.IO;
 using System.Linq;
 using System.Windows.Forms;
 
@@ -78,68 +74,9 @@ namespace Sdl.Community.DeepLMTProvider.DeepLTellMe
             }
         }
 
-        private static string GetDateTimeToString(DateTime dateTime)
-        {
-            var value = dateTime.Year +
-                        dateTime.Month.ToString().PadLeft(2, '0') +
-                        dateTime.Day.ToString().PadLeft(2, '0') +
-                        "-" +
-                        dateTime.Hour.ToString().PadLeft(2, '0') +
-                        dateTime.Minute.ToString().PadLeft(2, '0') +
-                        dateTime.Second.ToString().PadLeft(2, '0');
-            return value;
-        }
-
-        private static bool UserHasWritePermission(DirectoryInfo directoryInfo)
-        {
-            try
-            {
-                // Attempt to create a temporary file in the directory
-                string tempFilePath = Path.Combine(directoryInfo.FullName, Guid.NewGuid().ToString("N") + ".tmp");
-                using (FileStream fs = File.Create(tempFilePath)) { }
-
-                // Delete the temporary file
-                File.Delete(tempFilePath);
-
-                return true; // User has write permission
-            }
-            catch (UnauthorizedAccessException)
-            {
-                return false; // User does not have write permission
-            }
-        }
-
         private void ViewModel_ManageGlossaries()
         {
-            var glossaryReaderWriterFactory = new GlossaryReaderWriterFactory();
-            var messageService = new MessageService();
-
-            var glossariesWindowViewModel = new GlossariesWindowViewModel(new DeepLGlossaryClient(),
-                messageService, new GlossaryImportExportService(new DialogWrapper()),
-                new GlossaryReaderWriterService(glossaryReaderWriterFactory), new ProcessStarter(),
-                new EditGlossaryService());
-
-            var (success, glossaryWriter, _) = glossaryReaderWriterFactory.CreateGlossaryWriter(GlossaryReaderWriterService.Format.CSV);
-            if (!success) messageService.ShowWarning("Backup service could not be initialized.\nUse glossary manager carefully!", "Manage glossaries");
-
-            var backupPath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData),
-                "Trados AppStore",
-                $@"DeepLLogs\Backup");
-
-            if (!UserHasWritePermission(new DirectoryInfo(backupPath))) messageService.ShowWarning($"User not authorized to write to backup location {backupPath}");
-
-            glossariesWindowViewModel.ShouldBackUp += glossary =>
-            {
-                var backupDirectory = $@"{backupPath}\{GetDateTimeToString(DateTime.Now)}";
-                Directory.CreateDirectory(backupDirectory);
-
-                glossaryWriter.WriteGlossary(glossary,
-                    $@"{backupPath}\{GetDateTimeToString(DateTime.Now)}\{glossary.Name}_{glossary.SourceLanguage}-{glossary.TargetLanguage}.csv");
-            };
-
-            var glossariesWindow = new GlossariesWindow { DataContext = glossariesWindowViewModel };
-            glossariesWindow.ShowDialog();
-
+            GlossaryBackupService.ViewModel_ManageGlossaries();
             ViewModel.LoadLanguagePairSettings();
         }
     }
