+++ DeepLMTProvider/Sdl.Community.DeelLMTProvider/Studio/DeepLMtTranslationProviderWinFormsUI.cs
@@ -1,6 +1,4 @@
 ﻿using Sdl.Community.DeepLMTProvider.Client;
-using Sdl.Community.DeepLMTProvider.Helpers;
-using Sdl.Community.DeepLMTProvider.Helpers.GlossaryReadersWriters;
 using Sdl.Community.DeepLMTProvider.Model;
 using Sdl.Community.DeepLMTProvider.Service;
 using Sdl.Community.DeepLMTProvider.UI;
@@ -8,7 +6,6 @@ using Sdl.Community.DeepLMTProvider.ViewModel;
 using Sdl.LanguagePlatform.Core;
 using Sdl.LanguagePlatform.TranslationMemoryApi;
 using System;
-using System.IO;
 using System.Windows.Forms;
 using System.Windows.Forms.Integration;
 
@@ -116,37 +113,6 @@ namespace Sdl.Community.DeepLMTProvider.Studio
             return supportsProvider;
         }
 
-        private static string GetDateTimeToString(DateTime dateTime)
-        {
-            var value = dateTime.Year +
-                        dateTime.Month.ToString().PadLeft(2, '0') +
-                        dateTime.Day.ToString().PadLeft(2, '0') +
-                        "-" +
-                        dateTime.Hour.ToString().PadLeft(2, '0') +
-                        dateTime.Minute.ToString().PadLeft(2, '0') +
-                        dateTime.Second.ToString().PadLeft(2, '0');
-            return value;
-        }
-
-        private static bool UserHasWritePermission(DirectoryInfo directoryInfo)
-        {
-            try
-            {
-                // Attempt to create a temporary file in the directory
-                string tempFilePath = Path.Combine(directoryInfo.FullName, Guid.NewGuid().ToString("N") + ".tmp");
-                using (FileStream fs = File.Create(tempFilePath)) { }
-
-                // Delete the temporary file
-                File.Delete(tempFilePath);
-
-                return true; // User has write permission
-            }
-            catch (UnauthorizedAccessException)
-            {
-                return false; // User does not have write permission
-            }
-        }
-
         private TranslationProviderCredential GetCredentials(ITranslationProviderCredentialStore credentialStore, string uri)
         {
             var providerUri = new Uri(uri);
@@ -173,34 +139,7 @@ namespace Sdl.Community.DeepLMTProvider.Studio
 
         private void ViewModel_ManageGlossaries()
         {
-            var glossaryReaderWriterFactory = new GlossaryReaderWriterFactory();
-            var messageService = new MessageService();
-
-            var glossariesWindowViewModel = new GlossariesWindowViewModel(new DeepLGlossaryClient(),
-                messageService, new GlossaryImportExportService(new DialogWrapper()),
-                new GlossaryReaderWriterService(glossaryReaderWriterFactory), new ProcessStarter(), new EditGlossaryService());
-
-            var (success, glossaryWriter, _) = glossaryReaderWriterFactory.CreateGlossaryWriter(GlossaryReaderWriterService.Format.CSV);
-            if (!success) messageService.ShowWarning("Backup service could not be initialized.\nUse glossary manager carefully!", "Manage glossaries");
-
-            var backupPath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData),
-                "Trados AppStore",
-                $@"DeepLLogs\Backup");
-
-            if (!UserHasWritePermission(new DirectoryInfo(backupPath))) messageService.ShowWarning($"User not authorized to write to backup location {backupPath}");
-
-            glossariesWindowViewModel.ShouldBackUp += glossary =>
-            {
-                var backupDirectory = $@"{backupPath}\{GetDateTimeToString(DateTime.Now)}";
-                Directory.CreateDirectory(backupDirectory);
-
-                glossaryWriter.WriteGlossary(glossary,
-                    $@"{backupPath}\{GetDateTimeToString(DateTime.Now)}\{glossary.Name}_{glossary.SourceLanguage}-{glossary.TargetLanguage}.csv");
-            };
-
-            var glossariesWindow = new GlossariesWindow { DataContext = glossariesWindowViewModel };
-            glossariesWindow.ShowDialog();
-
+            GlossaryBackupService.ViewModel_ManageGlossaries();
             DeepLWindowViewModel.LoadLanguagePairSettings();
         }
     }
