+++ DeepLMTProvider/Sdl.Community.DeelLMTProvider/pluginpackage.manifest.xml
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <PluginPackage xmlns="http://www.sdl.com/Plugins/PluginPackage/1.0">
 	<PlugInName>DeepL Translation Provider</PlugInName>
-	<Version>6.1.2.5</Version>
+	<Version>6.1.2.6</Version>
 	<Description>This plugin provides machine translation results from the DeepL Translator.</Description>
 	<Author>Trados AppStore Team</Author>
 	<RequiredProduct name="TradosStudio" minversion="17.1" maxversion="17.9" />
